Create Invoice Program
Write a C++ program containing a class Invoice and a driver program called InvoiceDriver.cpp.
The class Invoice is used in a hardware store to represent an invoice for an item sold at the store.
An invoice class should include the following:
 A part number of type string
 A part description of type string
 A quantity of the item being purchased of type int
 A price per item of type int
 A class constructor that initializes the four data members
 An accessor (e.g., get member function) for each of the data member
 A mutator (i.e., a set member function) for each of the data member
 A member function named getInvoiceAmount that calculates the invoice amount (i.e.,
multiplies the quantity by the price per item, then it returns the amount as an int value
 The constructor and the mutators functions should take into account the following
conditions:
o i. If the quantity is not positive, it should be set to 0
o ii. If the price per item is not positive, it should be set to 0